Support for database snapshots.
Snapshots allow users to read specific versions of the database
at a given point in time. They can be read from a random database
state or they can be returned from calls that modify the database
to refer to a snapshot just after that update occurs but before
any other update.